Add reactive extensions to notification center. Remove notification manager layer and use center directly.
