+++ Core/Core/Notifications/UserNotificationCenterProtocol.swift
@@ -1,0 +1,67 @@
+//
+// This file is part of Canvas.
+// Copyright (C) 2024-present  Instructure, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+
+import Combine
+import UserNotifications
+
+public protocol UserNotificationCenterProtocol: AnyObject {
+    func requestAuthorization(options: UNAuthorizationOptions, completionHandler: @escaping (Bool, Error?) -> Void)
+    func add(_ request: UNNotificationRequest, withCompletionHandler completionHandler: ((Error?) -> Void)?)
+    func getPendingNotificationRequests(completionHandler: @escaping ([UNNotificationRequest]) -> Void)
+    func removePendingNotificationRequests(withIdentifiers identifiers: [String])
+    var delegate: UNUserNotificationCenterDelegate? { get set }
+}
+
+public extension UserNotificationCenterProtocol {
+
+    func requestAuthorization(options: UNAuthorizationOptions = []) -> Future<Void, Error> {
+        Future { [self] promise in
+            requestAuthorization(options: options) { granted, error in
+                if let error {
+                    return promise(.failure(error))
+                }
+
+                promise(granted ? .success(()) : .failure(NSError.instructureError("Denied")))
+            }
+        }
+    }
+
+    func add(_ request: UNNotificationRequest) -> Future<Void, Error> {
+        Future<Void, Error> { [self] promise in
+            add(request) { error in
+                if let error {
+                    Analytics.shared.logError(name: "Failed to schedule local notification",
+                                              reason: error.localizedDescription)
+                    promise(.failure(error))
+                } else {
+                    promise(.success(()))
+                }
+            }
+        }
+    }
+
+    func getPendingNotificationRequests() -> Future<[UNNotificationRequest], Never> {
+        Future { [self] promise in
+            getPendingNotificationRequests { notifications in
+                promise(.success(notifications))
+            }
+        }
+    }
+}
+
+extension UNUserNotificationCenter: UserNotificationCenterProtocol {}
